Fix the token verification when using a token that contains whitespace
Currently, the original token string it not stored anywhere. This token is decoded and those decoded values are stored. When you try to verify the signature, the header and payload are then re-encoded and checked against the signature. The issue with this is that the original formatting is lost when re-encoding the token. This change uses the original token, if present, to verify the signature against.